+++ packages/antd/src/templates/FieldHelpTemplate/index.tsx
@@ -1,6 +1,5 @@
-import { FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, helpId } from '@rjsf/utils';
+import { helpId, FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
 import { RichHelp } from '@rjsf/core';
-import Form from 'react-bootstrap/Form';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -11,13 +10,14 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help, uiSchema, registry, hasErrors } = props;
+  const { fieldPathId, help, uiSchema, registry } = props;
   if (!help) {
     return null;
   }
+
   return (
-    <Form.Text id={helpId(fieldPathId)} className={hasErrors ? 'text-danger' : 'text-muted'}>
+    <div id={helpId(fieldPathId)} className='help-block'>
       <RichHelp help={help} registry={registry} uiSchema={uiSchema} />
-    </Form.Text>
+    </div>
   );
 }