+++ packages/core/src/components/templates/FieldHelpTemplate.tsx
@@ -1,4 +1,5 @@
 import { helpId, FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import RichHelp from '../RichHelp';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -9,21 +10,14 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help } = props;
+  const { fieldPathId, help, uiSchema, registry } = props;
   if (!help) {
     return null;
   }
-  const id = helpId(fieldPathId);
-  if (typeof help === 'string') {
-    return (
-      <p id={id} className='help-block'>
-        {help}
-      </p>
-    );
-  }
+
   return (
-    <div id={id} className='help-block'>
-      {help}
+    <div id={helpId(fieldPathId)} className='help-block'>
+      <RichHelp help={help as string} registry={registry} uiSchema={uiSchema} />
     </div>
   );
 }