+++ packages/mantine/src/templates/FieldHelpTemplate.tsx
@@ -1,5 +1,6 @@
 import { helpId, FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
 import { Text } from '@mantine/core';
+import { RichHelp } from '@rjsf/core';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -10,17 +11,15 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help } = props;
+  const { fieldPathId, help, uiSchema, registry } = props;
 
   if (!help) {
     return null;
   }
 
-  const id = helpId(fieldPathId);
-
   return (
-    <Text id={id} size='sm' my='xs' c='dimmed'>
-      {help}
+    <Text id={helpId(fieldPathId)} size='sm' my='xs' c='dimmed'>
+      <RichHelp help={help} registry={registry} uiSchema={uiSchema} />
     </Text>
   );
 }