+++ packages/mui/src/FieldHelpTemplate/FieldHelpTemplate.tsx
@@ -1,5 +1,6 @@
-import FormHelperText from '@mui/material/FormHelperText';
+import { RichHelp } from '@rjsf/core';
 import { helpId, FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema } from '@rjsf/utils';
+import FormHelperText from '@mui/material/FormHelperText';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -10,14 +11,14 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help } = props;
+  const { fieldPathId, help, uiSchema, registry } = props;
   if (!help) {
     return null;
   }
-  const id = helpId(fieldPathId);
+
   return (
-    <FormHelperText component='div' id={id}>
-      {help}
+    <FormHelperText component='div' id={helpId(fieldPathId)} style={{ marginTop: '5px' }}>
+      <RichHelp help={help} registry={registry} uiSchema={uiSchema} />
     </FormHelperText>
   );
 }