+++ packages/semantic-ui/src/FieldHelpTemplate/FieldHelpTemplate.tsx
@@ -1,5 +1,6 @@
-import { Message } from 'semantic-ui-react';
 import { FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, helpId } from '@rjsf/utils';
+import { RichHelp } from '@rjsf/core';
+import { Message } from 'semantic-ui-react';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
  *
@@ -10,10 +11,13 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help } = props;
+  const { fieldPathId, help, uiSchema, registry } = props;
   if (help) {
-    const id = helpId(fieldPathId);
-    return <Message size='mini' info id={id} content={help} />;
+    return (
+      <Message size='mini' info id={helpId(fieldPathId)}>
+        <RichHelp help={help} registry={registry} uiSchema={uiSchema} />
+      </Message>
+    );
   }
   return null;
 }