+++ packages/shadcn/src/FieldHelpTemplate/FieldHelpTemplate.tsx
@@ -1,5 +1,5 @@
 import { FieldHelpProps, FormContextType, RJSFSchema, StrictRJSFSchema, helpId } from '@rjsf/utils';
-
+import { RichHelp } from '@rjsf/core';
 import { cn } from '../lib/utils';
 
 /** The `FieldHelpTemplate` component renders any help desired for a field
@@ -11,14 +11,17 @@ export default function FieldHelpTemplate<
   S extends StrictRJSFSchema = RJSFSchema,
   F extends FormContextType = any,
 >(props: FieldHelpProps<T, S, F>) {
-  const { fieldPathId, help, hasErrors } = props;
+  const { fieldPathId, help, uiSchema, registry, hasErrors } = props;
   if (!help) {
     return null;
   }
-  const id = helpId(fieldPathId);
+
   return (
-    <span className={cn('text-xs font-medium text-muted-foreground', { ' text-destructive': hasErrors })} id={id}>
-      {help}
+    <span
+      className={cn('text-xs font-medium text-muted-foreground', { 'text-destructive': hasErrors })}
+      id={helpId(fieldPathId)}
+    >
+      <RichHelp help={help} registry={registry} uiSchema={uiSchema} />
     </span>
   );
 }